refactor(worker): restructure modules and remove webhook tools
- Remove webhook tool execution (tools.rs, WebhookTool type)
- Simplify UnifiedToolExecutor to only use ToolRegistry
- Restructure worker modules:
  - runner_inprocess.rs + in_memory_workflows.rs → inprocess/
  - temporal/workflows.rs → temporal/workflows/ with trait-based registry
  - temporal/worker.rs → worker.rs (root level)
- Remove unused executor.rs (duplicate of InProcessRunner)
- Add WorkflowRegistry with builder pattern for dynamic workflow creation
- Update tool-execution.md spec to reflect builtin-only tools
