+++ crates/everruns-core/examples/capabilities.rs
@@ -92,9 +92,6 @@ impl OpenAiProvider {
             .iter()
             .map(|tool| {
                 let (name, description, parameters) = match tool {
-                    ToolDefinition::Webhook(webhook) => {
-                        (&webhook.name, &webhook.description, &webhook.parameters)
-                    }
                     ToolDefinition::Builtin(builtin) => {
                         (&builtin.name, &builtin.description, &builtin.parameters)
                     }
