+++ crates/everruns-core/examples/tool_calling.rs
@@ -86,9 +86,6 @@ impl OpenAiProvider {
             .iter()
             .map(|tool| {
                 let (name, description, parameters) = match tool {
-                    ToolDefinition::Webhook(webhook) => {
-                        (&webhook.name, &webhook.description, &webhook.parameters)
-                    }
                     ToolDefinition::Builtin(builtin) => {
                         (&builtin.name, &builtin.description, &builtin.parameters)
                     }
