+++ crates/everruns-worker/src/unified_tool_executor.rs
@@ -1,32 +1,20 @@
 // Unified Tool Executor
 //
-// This module provides a unified tool execution implementation that combines:
-// - ToolRegistry from agent-loop for built-in tools
-// - Webhook execution for external HTTP tools
-//
-// Both in-process and Temporal modes use this executor to ensure consistent
-// tool execution behavior.
+// This module provides tool execution using ToolRegistry from everruns-core.
+// All tools are registered at creation time and executed via the registry.
 
 use async_trait::async_trait;
 use everruns_contracts::tools::{ToolCall, ToolDefinition, ToolResult};
 use everruns_core::{traits::ToolExecutor, Result, ToolRegistry};
-use reqwest::Client;
 use std::sync::Arc;
 use tracing::{error, info};
 
-use crate::tools::{execute_tool as execute_webhook_tool, requires_approval};
-
-/// A unified tool executor that handles both built-in and webhook tools.
-///
-/// This executor:
-/// - Uses a `ToolRegistry` for built-in tools registered at creation time
-/// - Falls back to webhook execution for `ToolDefinition::Webhook` tools
-/// - For `ToolDefinition::Builtin` tools, looks up in the registry by name
+/// A tool executor that uses ToolRegistry for built-in tools.
 ///
 /// # Example
 ///
 /// ```ignore
-/// use everruns_agent_loop::{ToolRegistry, GetCurrentTime, EchoTool};
+/// use everruns_core::{ToolRegistry, GetCurrentTime, EchoTool};
 /// use everruns_worker::unified_tool_executor::UnifiedToolExecutor;
 ///
 /// // Create registry with built-in tools
@@ -35,7 +23,7 @@ use crate::tools::{execute_tool as execute_webhook_tool, requires_approval};
 ///     .tool(EchoTool)
 ///     .build();
 ///
-/// // Create unified executor
+/// // Create executor
 /// let executor = UnifiedToolExecutor::new(registry);
 ///
 /// // Use with AgentLoop
@@ -44,20 +32,17 @@ use crate::tools::{execute_tool as execute_webhook_tool, requires_approval};
 pub struct UnifiedToolExecutor {
     /// Registry of built-in tools
     registry: Arc<ToolRegistry>,
-    /// HTTP client for webhook calls
-    client: Client,
 }
 
 impl UnifiedToolExecutor {
-    /// Create a new unified tool executor with the given tool registry.
+    /// Create a new tool executor with the given tool registry.
     pub fn new(registry: ToolRegistry) -> Self {
         Self {
             registry: Arc::new(registry),
-            client: Client::new(),
         }
     }
 
-    /// Create a new unified tool executor with default built-in tools.
+    /// Create a new tool executor with default built-in tools.
     ///
     /// This includes:
     /// - `get_current_time`: Returns the current date and time
@@ -71,77 +56,9 @@ impl UnifiedToolExecutor {
         Self::new(registry)
     }
 
-    /// Create a new unified tool executor with an empty registry.
-    ///
-    /// Use this when you only need webhook tool execution.
-    pub fn webhook_only() -> Self {
-        Self::new(ToolRegistry::new())
-    }
-
     /// Get reference to the tool registry.
     pub fn registry(&self) -> &ToolRegistry {
         &self.registry
-    }
-
-    /// Check if a tool requires approval before execution.
-    pub fn tool_requires_approval(&self, tool_def: &ToolDefinition) -> bool {
-        requires_approval(tool_def)
-    }
-
-    /// Execute a built-in tool from the registry.
-    async fn execute_builtin(
-        &self,
-        tool_call: &ToolCall,
-        _builtin: &everruns_contracts::tools::BuiltinTool,
-    ) -> Result<ToolResult> {
-        // Look up the tool in the registry
-        if let Some(tool) = self.registry.get(&tool_call.name) {
-            info!(
-                tool_name = %tool_call.name,
-                tool_call_id = %tool_call.id,
-                "Executing built-in tool from registry"
-            );
-
-            let result = tool.execute(tool_call.arguments.clone()).await;
-            Ok(result.into_tool_result(&tool_call.id, &tool_call.name))
-        } else {
-            // Tool not found in registry
-            error!(
-                tool_name = %tool_call.name,
-                tool_call_id = %tool_call.id,
-                "Built-in tool not found in registry"
-            );
-
-            Ok(ToolResult {
-                tool_call_id: tool_call.id.clone(),
-                result: None,
-                error: Some(format!(
-                    "Built-in tool '{}' not found in registry",
-                    tool_call.name
-                )),
-            })
-        }
-    }
-
-    /// Execute a webhook tool.
-    async fn execute_webhook(
-        &self,
-        tool_call: &ToolCall,
-        tool_def: &ToolDefinition,
-    ) -> Result<ToolResult> {
-        info!(
-            tool_name = %tool_call.name,
-            tool_call_id = %tool_call.id,
-            "Executing webhook tool"
-        );
-
-        let result = execute_webhook_tool(tool_call, tool_def, &self.client).await;
-
-        Ok(ToolResult {
-            tool_call_id: result.tool_call_id,
-            result: result.result,
-            error: result.error,
-        })
     }
 }
 
@@ -153,10 +70,34 @@ impl Default for UnifiedToolExecutor {
 
 #[async_trait]
 impl ToolExecutor for UnifiedToolExecutor {
-    async fn execute(&self, tool_call: &ToolCall, tool_def: &ToolDefinition) -> Result<ToolResult> {
-        match tool_def {
-            ToolDefinition::Builtin(builtin) => self.execute_builtin(tool_call, builtin).await,
-            ToolDefinition::Webhook(_) => self.execute_webhook(tool_call, tool_def).await,
+    async fn execute(
+        &self,
+        tool_call: &ToolCall,
+        _tool_def: &ToolDefinition,
+    ) -> Result<ToolResult> {
+        // Look up the tool in the registry by name
+        if let Some(tool) = self.registry.get(&tool_call.name) {
+            info!(
+                tool_name = %tool_call.name,
+                tool_call_id = %tool_call.id,
+                "Executing tool from registry"
+            );
+
+            let result = tool.execute(tool_call.arguments.clone()).await;
+            Ok(result.into_tool_result(&tool_call.id, &tool_call.name))
+        } else {
+            // Tool not found in registry
+            error!(
+                tool_name = %tool_call.name,
+                tool_call_id = %tool_call.id,
+                "Tool not found in registry"
+            );
+
+            Ok(ToolResult {
+                tool_call_id: tool_call.id.clone(),
+                result: None,
+                error: Some(format!("Tool '{}' not found in registry", tool_call.name)),
+            })
         }
     }
 }
@@ -223,7 +164,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_builtin_tool_not_in_registry() {
+    async fn test_tool_not_in_registry() {
         let registry = ToolRegistry::new(); // Empty registry
         let executor = UnifiedToolExecutor::new(registry);
 
@@ -282,29 +223,5 @@ mod tests {
         assert!(executor.registry().has("get_current_time"));
         // Should have echo
         assert!(executor.registry().has("echo"));
-    }
-
-    #[test]
-    fn test_tool_requires_approval() {
-        let executor = UnifiedToolExecutor::webhook_only();
-
-        let auto_tool = ToolDefinition::Builtin(BuiltinTool {
-            name: "auto".to_string(),
-            description: "Auto tool".to_string(),
-            parameters: json!({}),
-            kind: BuiltinToolKind::HttpGet,
-            policy: ToolPolicy::Auto,
-        });
-
-        let approval_tool = ToolDefinition::Builtin(BuiltinTool {
-            name: "approval".to_string(),
-            description: "Approval tool".to_string(),
-            parameters: json!({}),
-            kind: BuiltinToolKind::HttpGet,
-            policy: ToolPolicy::RequiresApproval,
-        });
-
-        assert!(!executor.tool_requires_approval(&auto_tool));
-        assert!(executor.tool_requires_approval(&approval_tool));
     }
 }
